+++ app/src/main/java/com/cargopull/executor_driver/interactor/OrderCurrentCostGateway.java
@@ -12,8 +12,8 @@ public interface OrderCurrentCostGateway {
    * Запрашивает изменения цены выполняемого заказа.
    *
    * @param channelId - ID "канала", откуда брать баланс.
-   * @return {@link Flowable<Integer>} результат запроса.
+   * @return {@link Flowable<Long>} результат запроса.
    */
   @NonNull
-  Flowable<Integer> getOrderCurrentCost(@NonNull String channelId);
+  Flowable<Long> getOrderCurrentCost(@NonNull String channelId);
 }
