+++ app/src/main/java/com/cargopull/executor_driver/view/BalanceSummaryFragment.java
@@ -62,7 +62,7 @@ public class BalanceSummaryFragment extends BaseFragment implements BalanceViewA
   }
 
   @Override
-  public void showMainAccountAmount(int amount) {
+  public void showMainAccountAmount(long amount) {
     if (!getResources().getBoolean(R.bool.show_cents)) {
       amount = Math.round(amount / 100f);
     }
@@ -77,7 +77,7 @@ public class BalanceSummaryFragment extends BaseFragment implements BalanceViewA
   }
 
   @Override
-  public void showBonusAccountAmount(int amount) {
+  public void showBonusAccountAmount(long amount) {
 
   }
